netkit-0.17
bsd-finger:
    22-Jul-2000:
        Document ~/.pgpkey. (Herbert Xu, herbert@gondor.apana.org.au)

    18-Dec-1999:
        Add support for ~/.nofinger files.

linux-ftpd:
    8-Jul-2000:
        Fix exploitable misused printf-function calls.

netkit-base:
    24-Jul-2000:
        Don't use OPEN_MAX. (Olaf Kirch)

    23-Jul-2000:
        General audit for setuid/setgid/initgroups/etc. calls that didn't
          check for error.

    22-Jul-2000:
        Documented the virtual hosting stuff in inetd(8).
          (Michael Tokarev, mjt@tls.msk.ru)

        On systems where sleep() is implemented in terms of alarm(),
          inetd breaks because it tries to sleep while alarms are in
          progress, blocks SIGALRM, and other fun stuff. So, on libc5,
          reimplement sleep in terms of select to circumvent this.

        Fixed another file descriptor leak. (Pekka Savola,
          Pekka.Savola@netcore.fi)

        Added code that will hopefully restart inetd if it runs out of
          fds. It doesn't work terribly well yet, unfortunately.

    3-Jun-2000:
        Some minor fixes to ping.

    4-Feb-2000:
        Massive cleanup of ping.
        Also, added -w option (overall timeout).

    3-Feb-2000:
        Merge some ping fixes from Red Hat.

    20-Jan-2000:
        AF_UNIX sockets should be chmod 666'd after being bound.
          (Olaf Kirch, okir@caldera.de)

netkit-bootparamd:
    22-Jul-2000:
        Some cleanup. New scanf-free parsing code for bootparams file.
        Add bootparams(5) page (swiped from FreeBSD 4.0-RELEASE).

netkit-ftp:
    8-Jul-2000:
        Fix misused printf-function call (not %n-exploitable though).

netkit-ntalk:
    29-Jul-2000:
        Have talk scroll by screenfuls instead of single lines.
        Flip the escape-p/n and ctrl-p/n keys, so the other person's
          window scrolls using ctrl. The rationale is that one scrolls
          one's own screen much less often and ctrl-combinations are
          easier to type (and can be held down for repeat) - many
          keyboards do not have a shifting key that sends an escape
          prefix.

    22-Jul-2000:
        Fix minor nonexploitable overflow. (Found by "teso".)
        Fix input handling to not do silly things with 8-bit text input.
          (noted by several people)
        Hopefully fix a problem where talk spews trash on tty hangup.
        Also fix things so that talk doesn't exit until you say so - if
          the connection breaks, it won't let you type any more, but you
          can still scroll and read everything the other guy wrote.

    21-May-2000:
        Fix bug in otalk emulation. (Herbert Xu, herbert@gondor.apana.org.au)

    21-Dec-1999:
        Make talk handle window resize better.

netkit-routed:
    22-Jul-2000:
        Allow appending to an existing log file, as long as it's in
          the log directory. Also some cosmetic fixes. (Dennis Reichel,
          dennis@reichel.net)

        Also a bit of other cleanup and sanity checking.

    5-Jan-2000:
        Bug fix (mishandling of fork return codes). (Bill Nottingham,
          notting@redhat.com)
        Also correct getopt handling. (Noticed by Dirk von Suchodoletz,
          dirk@goe.net)

netkit-rpc:
    8-Jul-2000:
        Fix maybe-exploitable misused printf-function calls.

    12-Apr-2000:
        Improve the probing for cpp: now checks /usr/bin and /usr/libexec
          as well as /lib and /usr/ccs/lib.

netkit-rsh:
    23-Jul-2000:
        General audit for setuid/setgid/initgroups/etc. calls that didn't
          check for error.

    22-Jul-2000:
        Use getdtablesize() instead of OPEN_MAX. (Olaf Kirch, okir@caldera.de)

    5-Jan-2000:
        Fix to PAM handling. (Olaf Kirch, okir@caldera.de)
        Other PAM changes in rlogind. (Jeff Johnson, jbj@redhat.com,
          munged by yours truly.)

    17-Dec-1999:
        Fix bug in rshd (hangs forever with zombie offspring).
        Also fix problems with rlogind.
        To be posted as patch 1, making netkit-rsh-0.16.1.

netkit-rusers:
    22-Jul-2000:
        Fixed bug in rup: it wasn't able to print uptimes > 1 year.
          (Stephen G. Roseman, sgr0@lehigh.edu)
        Rearranged uptime printing logic in rup and added -s option
           to print times and uptimes in seconds for scripting.
        Least privilege patches (and some other fixes) for rusersd
           from Olaf Kirch.

    21-May-2000:
        Relax buffer paranoia a bit in rusersd to be able to send
          8-character usernames. Should be safe. (Rune Andresen,
          di990013@diku.dk)

netkit-rwho:
    22-Jul-2000:
        Add some minor additional paranoia.
        Changes from Olaf Kirch to run chrooted to /var/spool/rwho and
          then optionally drop root privileges.

    3-Jun-2000:
        Fix sizes of types in enclosed protocols/rwhod.h.

netkit-telnet:
    22-Jul-2000:
        Bug fixes for environment processing from Olaf Kirch. Also fixes
          privacy issue noticed by Steve Bellovin. Also fix a wrong
          assert().

    21-May-2000:
        Fix bug found by Herbert Xu (herbert@gondor.apana.org.au) - telnet
          was sending terminal type "(null)" as part of the terminal type
          list.

    12-Apr-2000:
        IPPROTO_IP is not a macro in Linux, so don't check it with #ifdef.
        Also, add initial experimental login wrapper, but don't make it
          part of the default build.

netkit-tftp:
    29-Jul-2000:
        Have tftp bind to the same address inetd was listening to,
          if it's not INADDR_ANY.

    22-Jul-2000:
        Fixes for file size limit bugs. Also, a few Solaris patches.
          (Ignacio Goyret, igoyret@acm.org)
        Switch to sigaction() and remove test for BSD signal semantics.

    28-Apr-2000:
        Bug fix: it seems if you started tftp and just pressed return,
          it would core... (Philipp Rumpf, prumpf@tux.org)

netkit-timed:
    23-Jul-2000:
        General audit for setuid/setgid/initgroups/etc. calls that didn't
          check for error.

+++ netkit-rwho/include/protocols/rwhod.h
@@ -31,7 +31,7 @@
  * SUCH DAMAGE.
  *
  *	from: @(#)rwhod.h	5.6 (Berkeley) 4/3/91
- *	$Id: rwhod.h,v 1.2 1996/07/15 21:41:36 dholland Exp $
+ *	$Id: rwhod.h,v 1.3 2000/06/03 16:43:06 dholland Exp $
  */
 
 #ifndef _RWHOD_H_
@@ -43,21 +43,21 @@
 struct	outmp {
 	char	out_line[8];		/* tty name */
 	char	out_name[8];		/* user id */
-	long	out_time;		/* time on */
+	int32_t	out_time;		/* time on */
 };
 
 struct	whod {
 	char	wd_vers;		/* protocol version # */
 	char	wd_type;		/* packet type, see below */
 	char	wd_pad[2];
-	int	wd_sendtime;		/* time stamp by sender */
-	int	wd_recvtime;		/* time stamp applied by receiver */
+	int32_t	wd_sendtime;		/* time stamp by sender */
+	int32_t	wd_recvtime;		/* time stamp applied by receiver */
 	char	wd_hostname[32];	/* hosts's name */
-	int	wd_loadav[3];		/* load average as in uptime */
-	int	wd_boottime;		/* time system booted */
+	int32_t	wd_loadav[3];		/* load average as in uptime */
+	int32_t	wd_boottime;		/* time system booted */
 	struct	whoent {
 		struct	outmp we_utmp;	/* active tty info */
-		int	we_idle;	/* tty idle time */
+		int32_t	we_idle;	/* tty idle time */
 	} wd_we[1024 / sizeof (struct whoent)];
 };
